% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaTepe.r
\docType{data}
\name{adaTepe}
\alias{adaTepe}
\title{Evidence from Bronze age gold mining}
\format{
A data frame with 177 rows and 4 variables:
\describe{
	\item{magSus}{mass specific magnetic susceptibility}
\item{logRatio1}{the log of the ratio of magnetic suspectibility measured 
    by anhysteretic remanent magnetization to frequency-dependent magnetic susceptibility}
\item{logRatio2}{the log of the ratio of magnetic suspectibility measured by anhystereticremanent
      magnetization to isothermal magnetization}
\item{sqrtHIRM}{the square-root of the hard-coercivity remanent magnetization}
}
}
\source{
Jordanova, N., Jordanova, D., Tcherkezova, E., Popov, H., Mokreva, A., 
	Georgiev, P., & Stoychev, R. (2020). Identification and classification 
   of archeological materials from Bronzeage gold mining site Ada Tepe (Bulgaria) 
using rock magnetism. \emph{Geochemistry, Geophysics, Geosystems}, vol.21, e2020GC009374.
}
\usage{
data(adaTepe)
}
\description{
In a study archaeologists collected rock samples that were linked with 
Bronze age gold mining. All the samples were taken from a dig at Ada Tepe 
in South-Eastern Bulgaria. In this datset there are just four of the variables 
that the archaeologists measured for each sample, all relating to the 
magnetic properties of the sample.
}
\keyword{datasets}
