% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brexit.r
\docType{data}
\name{brexit}
\alias{brexit}
\title{Voter behaviour in the Brexit referendum}
\format{
A data frame with 237 rows and 6 variables:
\describe{
 \item{leave}{percentage who voted `Leave' in the referendum}
 \item{age}{percentage who are in the age group 30 to 44 years old (based on
the 2001 census)}
 \item{income}{mean hourly pay in pounds sterling (based on the 2005 Annual Survey of
Hours and Earnings)}
\item{migrant}{percentage who are EU migrants (based on the 2001
census)}
\item{satisfaction}{mean life satisfaction score (from the Annual
Population Survey 2015)}
\item{noQual}{ percentage with no educational qualification (based on the
2001 census)}
}
}
\source{
Becker, S. O., Fetzer, T. and Novy, D. (2017). Who voted for Brexit? A comprehensive 
district-level analysis. \emph{Economic Policy}, 32, pp. 601-650.
}
\usage{
data(brexit)
}
\description{
The data considered in this dataset are a subset of the data collected about the UK referendum 
held on 23 June 2016. This referendum asked the UK electorate to vote on whether the UK should 
remain in the European Union (EU) or leave the EU: The dataset gives vote shares at the local 
authority level.
}
\keyword{datasets}
