% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carPrices.r
\docType{data}
\name{carPrices}
\alias{carPrices}
\title{The value of used General Motors cars}
\format{
A data frame with 793 rows and 4 variables:
\describe{
 \item{price}{suggested retail price of each car (in \$);}
 \item{mileage}{number of miles the car has been driven (in thousands of
miles)}
 \item{make}{manufacturer of the car such as Saturn, Pontiac, and
Chevrolet}
 \item{liter}{measure of engine size}
}
}
\source{
Kuiper, S. (2008). Introduction to Multiple Regression: How
 Much Is Your Car Worth? \emph{Journal of Statistics Education}, 16, DOI:
10.1080/10691898.2008.11889579.
}
\usage{
data(carPrices)
}
\description{
General Motors is an American multinational corporation that
encompasses different makes of cars. Data are available for several
hundreds of used General Motors cars in excellent condition. All of the
cars in the dataset were less than one year old when priced in 2005.
The condition of a car is thought to greatly affect its price and there
was interest in developing a multiple regression model to predict a
car's price based on a variety of characteristics such as its mileage,
make, model, engine size and interior style.
This dataset considered in this unit is a subset of 800
observations from these data.
}
\keyword{datasets}
