% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clothingFirms.r
\docType{data}
\name{clothingFirms}
\alias{clothingFirms}
\title{European firms producing textiles and footwear}
\format{
A data frame with 20096 rows and 5 variables:
\describe{
\item{segment}{A SIC (standard industry classification) code indicating the precise 
		segment of industry that the firm is engaged in}
\item{year}{the year that the observation was made}
\item{production}{the production of the firm represented by its gross sales}
\item{labour}{the labour input to the firm's production, represented by the 
	number of employees}
\item{capital}{the capital input to the firm's production, represented by 
	the value of its fixed assets in the year of observation}
}
}
\source{
Bureau van Dijk 
\url{https://www.eui.eu/Research/Library/ResearchGuides/Economics/Statistics/DataPortal/AmadeusBvD}
}
\usage{
data(clothingFirms)
}
\description{
A cross-sectional dataset of 20096 European firms engaged in the industry sector 
involved with the manufacture of textiles, wearing apparel and footwear.
}
\keyword{datasets}
