% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan.r
\name{dbscan}
\alias{dbscan}
\title{A function to implement the clustering algorithm DBScan}
\usage{
dbscan(dataset, minpts = 4, epsilon = 3, method = "euclidean")
}
\arguments{
\item{dataset}{a data frame, matrix or vector containing the observations}

\item{minpts}{the numbers of observations (including itself) that an observation needs 
to be close to an interior point}

\item{epsilon}{a parameter giving the maximum separation between two `close' observations}

\item{method}{the dissimilarity measire to be used. Note this must be one if the options 
available in \code{dist}}
}
\value{
A vector giving the cluster allocations for each observation.
with "-1" for outliers.
}
\description{
A function to implement the clustering algorithm DBScan. The output gives the allocation for each
observation, with -1 denoting when an observation is deemed to be an outlier.
}
\examples{
X <- c(10,8,9,2,1,0,-4)
clusters <- dbscan(X, minpts=2,epsilon=3)

}
\author{
Karen Vines
}
\keyword{dbscan}
