% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dist.r
\name{dissimPlot}
\alias{dissimPlot}
\title{A plot function for dist objects}
\usage{
dissimPlot(distobj, clusters = NULL, names = NULL, max.dist = NULL)
}
\arguments{
\item{distobj}{the dist object to be plotted}

\item{clusters}{a vector giving the cluster allocation for the data points used to create distobj}

\item{names}{a vector giving the names of the clusters}

\item{max.dist}{a value giving the assumed maximum dissimilarity in the matrix}
}
\value{
A plot of the dissimilarity matrix.
}
\description{
A function to produce plots of dissimilarity matrices. It assumes that the input is in the 
form of a dist object (produced by the function dist()). There is the option to order the 
data points by a clusters given by a clustering algorithm.
}
\examples{
X <- c(10,8,9,2,5,0)
distX <- dist(X)
# Plot of the dissimilarity matrix
dissimPlot(distX)
# Plot of the dissimilarity ordering by cluster
dissimPlot(distX, clusters=c(1,1,1,2,2,2))

}
\seealso{
See also \code{\link{dist}}.
}
\author{
Karen Vines
}
\keyword{plot}
