% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durbinWatson.r
\name{durbinWatson}
\alias{durbinWatson}
\title{A function to calculate the Durbin-Watson statistic}
\usage{
durbinWatson(lmFit)
}
\arguments{
\item{lmFit}{an object obtained after fitting a linear model}
}
\value{
The Durbin-Watson statistic
}
\description{
A function to calculate the DDurbin-Watson statistic for a set of residuals from a linear model.
}
\author{
Karen Vines
}
