% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employmentRate.r
\docType{data}
\name{employmentRate}
\alias{employmentRate}
\title{Employment rates}
\format{
A data frame with 296 rows and 4 variables:
\describe{
 \item{country}{the name of the country}
 \item{rate}{the 2019 employment rate (as a percentage) for people
educated to a Bachelor's degree or equivalent level}
 \item{gender}{the gender the inidividual identifies with taking the two possible values male and female}
 \item{age}{the age groupings (in years), taking the
 values 25 to 34, 35 to 44, 45 to 54 and 55 to 64}
}
}
\source{
\url{https://stats.oecd.org/}
}
\usage{
data(employmentRate)
}
\description{
This dataset contains data on the 2019 employment
   rates in 46 countries for people educated to a Bachelor's degree or
equivalent level, broken down by gender and age grouping, collected by the OECD.
}
\keyword{datasets}
