% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/europeanCompanies.r
\docType{data}
\name{europeanCompanies}
\alias{europeanCompanies}
\title{Public and private companies across Europe}
\format{
A data frame with 270 rows and 4 variables:
\describe{
   \item{country}{a categorical variable taking the values GB (for United
Kingdom of Great Britain and Northern Ireland), FR (for France)
and DE (for Germany)}
   \item{product}{a categorical variable giving the industrial classification
linked to what a company produces taking the coded following
values: 10 (for manufacture of food products), 20 (for manufacture
of chemicals and chemical products), 22 (for manufacture of rubber
and plastic products), 26 (for manufacture of computer, electronic
and optimal products), 27 (for manufacture of electrical equipment)}
   \item{averageWage}{average wage of an employee in 2018 (calculated by
taking the cost of the employees for that company in 2018 (in
thousands of Euro) and dividing by the number of employees in
2018)}
   \item{resAndDev}{a binary variable taking the values 1 (if the company
participates in research and development) and 0 (if the company
doesn't participate in research and development)}
}
}
\source{
Bureau van Dijk 
\url{https://www.eui.eu/Research/Library/ResearchGuides/Economics/Statistics/DataPortal/AmadeusBvD}
}
\usage{
data(europeanCompanies)
}
\description{
The Amadeus data is taken from a database which contains
comprehensive information on around 21 million public and private
companies across Europe. There are around 21
million companies in the database with information on financial
indicators, stock prices, directors and many other attributes. The
data can be used to explore companies' competitiveness, economic
integration, applied microeconomics, business cycles, economic
geography and corporate finance.
}
\keyword{datasets}
