% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook.r
\docType{data}
\name{facebook}
\alias{facebook}
\title{Patterns of social media usage}
\format{
A data frame with 135 rows and 3 variables:
\describe{
\item{friends}{the number of Facebook `friends' each person has}
\item{age}{the person's age in years}
\item{visits}{the number of visits a person makes in a day to the Facebook website}
}
}
\usage{
data(facebook)
}
\description{
This dataset contain a subset of data from a recent study on patterns of social media 
usage between adults. It corresponds to 135 Facebook users.
}
\keyword{datasets}
