% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasten.r
\docType{data}
\name{fasten}
\alias{fasten}
\title{Strength of alloy fasteners}
\format{
A data frame with 10 rows and 3 variables:
\describe{
  \item{load}{pressure load (in pounds per square inch)}
  \item{failed}{number in sample that failed}
  \item{samsize}{sample size}
}
}
\source{
Montgomery, D. C. and Peck, E. A. (1982) \emph{Introduction to Linear Regression Analysis}, 
 John Wiley, New York.
}
\usage{
data(fasten)
}
\description{
Ten samples of alloy fasteners used in aircraft were tested, each sample at a different pressure load.
}
\keyword{datasets}
