% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMresidplot.r
\name{glmResidPlot}
\alias{glmResidPlot}
\title{A function to produce residual plots for generalised linear models}
\usage{
glmResidPlot(glmobj, plots = c(1, 2, 3, 4))
}
\arguments{
\item{glmobj}{an object produced by the fitting of a model using the function glm()}

\item{plots}{plots to be produced where 1 for a plot of residuals versus 
transformed fitted values; 
2 for an index plot of residuals; 
3 for an index plot of squared residuals; and
4 for normal plot of residuals}
}
\value{
A plot of the silhouette statistics.
}
\description{
A function to produce residual plots for generalised linear models
}
\examples{
y <- c(10,8,9,2,5,0)
x <- c(1,2,3,4,5,7)
glmfit <- glm(y~x, family="poisson")
glmResidPlot(glmfit)

}
\seealso{
{\code{\link{glm}}}
}
\author{
Karen Vines
}
\keyword{glm}
