% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hausman.r
\name{hausman}
\alias{hausman}
\title{A function to carry out the Hausman test}
\usage{
hausman(feObj, reObj)
}
\arguments{
\item{feObj}{an object obtained after using the LSDV estimator}

\item{reObj}{an object obtained after using the random effects estimator.}
}
\value{
The results from carrying out the Hausman test
}
\description{
A function to carry out the Hausman test. The null hypothesis is that
the estimates from a fixed effects estimator and a random effects estimator are consistent.
It is assumed that as a fixed effects estimator, the LSDV estimator is used and for the random
effects estimator, the results from \code{lme()} are used.
}
\author{
Karen Vines
}
