% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importExport.r
\docType{data}
\name{importExport}
\alias{importExport}
\title{UK imports and exports}
\format{
A data frame with 100 rows and 5 variables:
\describe{
\item{year}{the year the observation relates to}
\item{quarter}{the quarter the observation relates to}
\item{exports}{value of real exports}
\item{income}{value of national income}
\item{imports}{value of real imports}
}
}
\source{
Eurostat
}
\usage{
data(importExport)
}
\description{
A time series dataset containing quarterly data for imports, exports and 
income for the UK from 1995 to 2019.
}
\keyword{datasets}
