% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagoon.r
\docType{data}
\name{lagoon}
\alias{lagoon}
\title{The salinity of water in the Bimini Lagoon}
\format{
A data frame with 24 observations and 2 variables:
\describe{
\item{salinity}{salinity (parts per thousand)}
\item{waterMass}{water mass (`1', `2', or `3')}
}
}
\source{
Till, R. (1974) \emph{Statistical Methods for the Earth Scientist}, London, Macmillan.
}
\usage{
data(lagoon)
}
\description{
Data from a study of salinity values for samples from three separate
water masses in the Bimini Lagoon, Bahamas.
}
\keyword{datasets}
