% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M348-package.R
\docType{package}
\name{M348-package}
\alias{M348}
\alias{M348-package}
\title{M348: Bespoke data files and functions for use with M348}
\description{
Bespoke data files and functions for use with M348.
}
\details{
Collection of datasets and functions for use when studying M348 Applied statistical modelling
}
\keyword{internal}
