% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parking.r
\docType{data}
\name{parking}
\alias{parking}
\title{Occupancy of a car park}
\format{
{A data frame with 73 observations and 3 variables:}
\describe{
   \item{calendarDate}{the date which the observation refers to}
   \item{day}{the day of the week}
   \item{occupancy}{the number of cars parked}
}
}
\source{
Stolfi, D.H., Alba, E. and Yao, X. Predicting Car Park Occupancy Rates
in Smart Cities. In: \emph{Smart Cities: Second International Conference},
Smart-CT 2017, Malaga, Spain, June 14-16, 2017, pp. 107--17.
}
\usage{
data(parking)
}
\description{
The number of cars parked in various Birmingham car parks in the
autumn 2016 were recorded by a car park company. 
This dataset gives data for one of these car parks, the Broad Street car park, at around noon.
}
\keyword{datasets}
