% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pea.r
\docType{data}
\name{pea}
\alias{pea}
\title{The effect of sugars on pea growth}
\format{
A data frame with 50 observations and 2 variables:
\describe{
 \item{length}{length of pea section (in coded units)}
 \item{treatment}{treatment used (`none', `glucose', `fructose', `gluc + fruct' or `sucrose')}
}
}
\source{
Sokal, R. R. and Rohlf, F. J. (1981) \emph{Biometry}, 2nd edition, San Francisco, W. H. Freeman.
}
\usage{
data(pea)
}
\description{
Data from an experiment in plant physiology recorded the lengths of
pea sections. The pea sections were grown in an appropriate tissue
culture in the presence of each of five different treatments. The aim
was to test the effects of different sugars on pea growth. The lengths
were given in 'coded units', which implies that, while meaning
something to the experimenter, the units do not mean anything to us.
The five treatments differed only in the presence or absence of
different sugars: one treatment contained no sugars, whereas the
others contained different quantities of different sugars.
For each of the five treatments, data for 10 pea sections were recorded.
}
\keyword{datasets}
