% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peru.r
\docType{data}
\name{peru}
\alias{peru}
\title{Blood pressue of Indigenous Peruvians}
\format{
A data frame with 38 rows and 9 variables:
\describe{
\item{age}{age (years)}
\item{years}{years since migration}
\item{weight}{weight (kg)}
\item{height}{height (mm)}
\item{chin}{chin skin fold (mm)}
\item{forearm}{forearm skin fold (mm)}
\item{calf}{calf skin fold (mm)}
\item{pulse}{pulse rate (beats per minute)}
\item{sbp}{systolic blood pressure}
}
}
\source{
Baker, P. T. and Beall, C. M. (1982) `The biology and health of Andean 
migrants: a case study in south coastal Peru', \emph{Mountain Research and
Development}, \bold{2}
}
\usage{
data(peru)
}
\description{
A dataset of measurements
on 38 Peruvian indians. All these individuals were males, over 21.
They were all born at high altitude in the Peruvian Andes, and their parents were
born at high altitude too. However, all these people had since moved
 into the mainstream of Peruvian society at a much lower altitude.
}
\keyword{datasets}
