% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcoaster.r
\docType{data}
\name{rcoaster}
\alias{rcoaster}
\title{Attributes of roller coasters}
\format{
A data frame with 236 rows and 3 variables:
\describe{
 \item{speed}{maximum speed in miles per hour}
 \item{height}{height of the roller coaster, measured in feet}
 \item{length}{length of the roller coaster, also measured in feet}
}
}
\source{
\url{rcdb.com}
}
\usage{
data(rcoaster)
}
\description{
Data on roller coasters around the world. Newly opened roller coasters are 
added to the database on a regular basis.  This dataset contains data on a 
subset of the data kept at rcdb.com.
}
\keyword{datasets}
