% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheepLitters.r
\docType{data}
\name{sheepLitters}
\alias{sheepLitters}
\title{Litters of sheep}
\format{
A data frame with 36 rows and 4 variables:
\describe{
 \item{count}{the number of ewes}
 \item{litterSize}{the number of lambs born in the litter, taking the values 0, 1, 2 and '>= 3'}
 \item{breed}{the breed of the ewe, taking the values a, b and c.}
 \item{farm}{the farm where the birth took place, taking the values 1, 2 and 3}
}
}
\source{
Mead, Curnow and Hasted (2002), Statistical methods in agriculture and experimental biology, Chapman and Hall/CRC Press, London.
}
\usage{
data(sheepLitters)
}
\description{
Agricultural researchers were interested in investigating the
relationships between the size of litters of lambs, the breed of ewe
giving birth to the lambs (from three possible breeds), and the farm
 where the ewe gave birth (from three possible farms).
}
\keyword{datasets}
