% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{silhouette}
\alias{silhouette}
\title{Calculation of silhouette statistics}
\usage{
silhouette(data, clusters, method = "euclidean")
}
\arguments{
\item{data}{a data frame, matrix or vector}

\item{clusters}{a vector giving the cluster allocation for the data points used to create distobj}

\item{method}{the dissimilarity measure to be used}
}
\value{
A vector giving the silhouette statistic for each observation.
}
\description{
A function to calculate silhouette statistics for a cluster allocation. For data point \eqn{i}, 
the silhouette statistic is defined as  
\deqn{s_i = \frac{b_i - a_i}{\max{(a_i,b_i)}}}
where \eqn{a_i} is the average dissimilarity between data point \eqn{i} and the other data points 
in the same cluster and \eqn{b_i} is the average dissimarity between data point \eqn{i} and the 
data points in the next nearest cluster.
}
\examples{
X <- c(10,8,9,2,5,0)
silAns <- silhouette(X, clusters=c(1,1,1,2,2,2))
plot(silAns)
summary(silAns)

}
\seealso{
\code{\link{plot.silhouette}}, \code{\link{summary.silhouette}}.
}
\author{
Karen Vines
}
