% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatedPanel.r
\docType{data}
\name{simulatedPanel}
\alias{simulatedPanel}
\title{Aggregate income and consumption}
\format{
A data frame with 60 rows and 4 variables:
\describe{
\item{country}{the country the observation relates to}
\item{year}{the time period the observation relates to}
\item{income}{aggregate income of the country}
\item{consumption}{aggregate consumption of the country}
}
}
\usage{
data(simulatedPanel)
}
\description{
This fictional dataset includes observations of aggregate income and aggregate 
consumption for five hypothetical countries over twelve time periods.
}
\keyword{datasets}
