% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukSurveyTableEmploymentGender.r
\docType{data}
\name{ukSurveyEmploymentGender}
\alias{ukSurveyEmploymentGender}
\title{Employment and Gender in a Living Costs and Food Survey}
\format{
A data frame with 8 rows and 3 variables:
\describe{
\item{count}{the number of household reference people}
\item{employment}{the type of employment of the household reference
 person (HRP), taking the values full-time, part-time, unemployed
 and inactive (that is, retired or not working and not looking for
 work)}
	\item{gender}{the gender that the HRP identifies with, taking the values
male and female}
}
}
\source{
Cathie Marsh Institute for Social Research (2019) 
 `Living Costs and Food Survey, 2013: Unrestricted Access Teaching Dataset'. 
  2nd edn. Available at: \url{https://doi.org/10.5255/UKDA-SN-7932-2}.
}
\usage{
data(ukSurveyEmploymentGender)
}
\description{
The data relating to employment and gender from \code{UKSurvey} given in 
a tabulated form.
}
\keyword{datasets}
