% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usConsumption.r
\docType{data}
\name{usConsumption}
\alias{usConsumption}
\title{GDP and consumption in the USA}
\format{
A data frame with 88 rows and 4 variables:
\describe{
	\item{year}{year the observation relates to}
	\item{quarter}{quarter the observation relates to, taking values 1, 2, 3 and 4} 
	\item{gdp}{quarterly US GDP in billions of constant 1987 US dollars}
	\item{pce}{quarterly personal consumption expenditure for the US economy in billions of constant 1987 US dollars}
}
}
\usage{
data(usConsumption)
}
\description{
This dataset contains quarterly time series data for GDP and personal 
consumption expenditure for the US economy from quarter 1 of 1970 to quarter 4 
of 1991, a total of 88 observations.
}
\keyword{datasets}
