% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wageEducation2005.r
\docType{data}
\name{wageEducation2005}
\alias{wageEducation2005}
\title{Wages and education in 2005}
\format{
A data frame with 775 observations and 5 variables:
\describe{
\item{id}{a numerical identifier for the individual}
	\item{age}{age (years)}
	\item{experience}{number of years in work}
\item{education}{years of education}
	\item{hourlyWage}{hourly wage (in wage units)}
}
}
\usage{
data(wageEducation2005)
}
\description{
This dataset consists of a single year, 2005, taken from the larger panel
\code{wageEducation}. It is a fictional dataset with observations of wage and 
experience data for 775 individuals from a fictional country recorded in 2005.
}
\keyword{datasets}
